quete express 8, les jetons d'auth

+++ userHandlers.js
@@ -90,39 +90,68 @@ const updateUser = (req, res) => {
   const id = parseInt(req.params.id);
   const { firstname, lastname, email, city, language } = req.body;
 
-  database
-    .query(
-      "update movies set firstname = ?, lastname = ?, email = ?, city = ?, language = ? where id = ?",
-      [firstname, lastname, email, city, language, id]
-    )
-    .then(([result]) => {
-      if (result.affectedRows === 0) {
-        res.status(404).send("Not Found");
-      } else {
-        res.sendStatus(204);
-      }
-    })
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send("Error editing the user");
-    });
+
+  if (id != parseInt(req.payload.sub)){
+    res.sendStatus(403);
+  } else 
+    { database
+      .query(
+        "update users set firstname = ?, lastname = ?, email = ?, city = ?, language = ? where id = ?",
+        [firstname, lastname, email, city, language, id]
+      )
+      .then(([result]) => {
+        if (result.affectedRows === 0) {
+          res.status(404).send("Not Found");
+        } else {
+          res.sendStatus(204);
+        }
+      })
+      .catch((err) => {
+        console.error(err);
+        res.status(500).send("Error editing the user");
+      });
+    }
 };
 
 const deleteUser = (req, res) => {
   const id = parseInt(req.params.id);
 
+  if (id != parseInt(req.payload.sub)){
+    res.sendStatus(403);
+  } else 
+    {
+    database
+      .query("delete from users where id = ?", [id])
+      .then(([result]) => {
+        if (result.affectedRows === 0) {
+          res.status(404).send("Not Found");
+        } else {
+          res.sendStatus(204);
+        }
+      })
+      .catch((err) => {
+        console.error(err);
+        res.status(500).send("Error deleting the user");
+      });
+    }
+};
+
+const getUserByEmailWithPasswordAndPassToNext = (req, res, next) => {
+  const email = req.body.email;
+
   database
-    .query("delete from users where id = ?", [id])
-    .then(([result]) => {
-      if (result.affectedRows === 0) {
-        res.status(404).send("Not Found");
+    .query("select email, hashedPassword, id from users where email = ?", [email])
+    .then(([users]) => {
+      if (users[0] != null) {
+        req.user=users[0];
+        next();
       } else {
-        res.sendStatus(204);
+        res.status(401).send("Not Found");
       }
     })
     .catch((err) => {
       console.error(err);
-      res.status(500).send("Error deleting the user");
+      res.status(500).send("Error retrieving data from database");
     });
 };
 
@@ -132,4 +161,5 @@ module.exports = {
   postUser,
   updateUser,
   deleteUser,
+  getUserByEmailWithPasswordAndPassToNext,
 };
